+++ package.json
@@ -143,6 +143,7 @@
     "get-tsconfig": "^4.7.5",
     "rollup": "^4.24.2",
     "rollup-plugin-esbuild": "^6.1.1",
-    "rollup-plugin-postcss": "^4.0.2"
+    "rollup-plugin-postcss": "^4.0.2",
+    "vite-tsconfig-paths": "^4.2.0"
   }
 }
