+++ packages/date-adapters/package.json
@@ -38,6 +38,6 @@
     "provenance": true
   },
   "scripts": {
-    "build": "yarn node ../../scripts/build.mjs"
+    "build": "yarn node ./scripts/build.mjs"
   }
 }
