+++ packages/date-adapters/src/dayjs-adapter/index.ts
@@ -1,7 +1,12 @@
-import defaultMoment, { type Moment } from "moment";
+import defaultDayjs, { type Dayjs } from "dayjs";
+import customParseFormat from "dayjs/plugin/customParseFormat";
+import localeData from "dayjs/plugin/localeData";
+import timezone from "dayjs/plugin/timezone";
+import utc from "dayjs/plugin/utc";
+import weekday from "dayjs/plugin/weekday";
 import {
   type AdapterOptions,
-  DateDetailErrorEnum,
+  DateDetailError,
   type ParserResult,
   type RecommendedFormats,
   type SaltDateAdapter,
@@ -9,24 +14,32 @@ import {
   type Timezone,
 } from "../types";
 
-declare module "../types" {
-  export interface DateFrameworkTypeMap {
-    moment: Moment;
+type Constructor = {
+  (...args: Parameters<typeof defaultDayjs>): Dayjs;
+  tz?: (value: Parameters<typeof defaultDayjs>[0], timezone: string) => Dayjs;
+  utc?: (value?: Parameters<typeof defaultDayjs>[0]) => Dayjs;
+};
+
+declare module "@salt-ds/date-adapters" {
+  interface DateFrameworkTypeMap {
+    dayjs: Dayjs;
   }
 }
 
+defaultDayjs.extend(utc);
+defaultDayjs.extend(timezone);
+defaultDayjs.extend(weekday);
+defaultDayjs.extend(localeData);
+
 /**
- * Adapter for Moment.js library, implementing the SaltDateAdapter interface.
- * Provides methods for date manipulation and formatting using Moment.js.
- * Salt provides a Moment adapter to aid migration to a maintained library.
- *
- * @deprecated Moment date library has been deprecated by its maintainers since September 2020, consider migration to a maintained OSS library.
+ * Adapter for Day.js library, implementing the SaltDateAdapter interface.
+ * Provides methods for date manipulation and formatting using Day.js.
  */
-export class AdapterMoment implements SaltDateAdapter<Moment, string> {
+export class AdapterDayjs implements SaltDateAdapter<Dayjs, string> {
   /**
-   * The Moment.js instance used for date operations.
+   * The Day.js instance used for date operations.
    */
-  public moment: typeof defaultMoment;
+  public dayjs: Constructor;
 
   /**
    * The locale used for date formatting.
@@ -36,146 +49,139 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
   /**
    * The name of the date library.
    */
-  public lib = "moment";
+  public lib = "dayjs";
 
   /**
-   * Creates an instance of AdapterMoment.
-   * @param options - Adapter options including locale and instance.
+   * Creates an instance of AdapterDayjs.
+   * @param options - Adapter options including locale.
+   * @param instance - The Day.js instance to use.
    */
-  constructor({
-    locale,
-    instance,
-  }: AdapterOptions<string, typeof defaultMoment> = {}) {
-    this.moment = instance || defaultMoment;
+  constructor(
+    { locale }: AdapterOptions<string, typeof defaultDayjs> = {},
+    instance?: Constructor,
+  ) {
     this.locale = locale || "en";
+    this.dayjs = instance || defaultDayjs;
+
+    // Allow user customization
+    if (!instance) {
+      defaultDayjs.extend(customParseFormat);
+    }
   }
 
   /**
-   * Checks if the timezone plugin is available.
-   * @returns True if the timezone plugin is available, false otherwise.
+   * Type guard for Dayjs object
+   * @param date
+   * @private
    */
-  private hasTimezonePlugin = () =>
-    typeof (this.moment as any).tz !== "undefined";
+  private isDayjs(date: any): date is Dayjs {
+    return date instanceof this.dayjs().constructor;
+  }
 
   /**
-   * Creates a Moment.js date object in the system timezone.
+   * Creates a Day.js date object in the system timezone.
    * @param value - The date string to parse.
    * @param locale - The locale to use for parsing.
-   * @returns The parsed Moment.js date object.
+   * @returns The parsed Day.js date object.
    */
   private createSystemDate = (
     value: string | undefined,
     locale?: string,
-  ): Moment => {
-    const parsedValue = this.moment(value).local();
-    if (this.locale === undefined && locale === undefined) {
-      return parsedValue;
-    }
-
-    return parsedValue.locale(locale ?? this.locale);
+  ): Dayjs => {
+    return this.dayjs(value).locale(locale ?? this.locale);
   };
 
   /**
-   * Creates a Moment.js date object in UTC.
+   * Creates a Day.js date object in UTC.
    * @param value - The date string to parse.
    * @param locale - The locale to use for parsing.
-   * @returns The parsed Moment.js date object.
+   * @returns The parsed Day.js date object.
+   * @throws Error if the UTC plugin is missing.
    */
-  private createUTCDate = (value: string, locale?: string): Moment => {
-    const parsedValue = this.moment.utc(value);
-    if (this.locale === undefined && locale === undefined) {
-      return parsedValue;
+  private createUTCDate = (
+    value: string | undefined,
+    locale?: string,
+  ): Dayjs => {
+    if (!this.dayjs.utc) {
+      throw new Error("Salt Day.js adapter: missing UTC plugin");
     }
-
-    return parsedValue.locale(locale ?? this.locale);
+    return this.dayjs.utc(value).locale(locale ?? this.locale);
   };
 
   /**
-   * Creates a Moment.js date object in a specified timezone.
+   * Creates a Day.js date object in a specified timezone.
    * @param value - The date string to parse.
    * @param timezone - The timezone to use.
    * @param locale - The locale to use for parsing.
-   * @returns The parsed Moment.js date object.
+   * @returns The parsed Day.js date object.
    * @throws Error if the timezone plugin is missing.
    */
   private createTZDate = (
-    value: string,
+    value: string | undefined,
     timezone: Timezone,
     locale?: string,
-  ): Moment => {
-    if (!this.hasTimezonePlugin()) {
-      throw new Error("Salt moment adapter: missing timezone plugin");
+  ): Dayjs => {
+    if (!this.dayjs.tz) {
+      throw new Error("Salt Day.js adapter: missing timezone plugin");
     }
-    const parsedValue =
-      timezone === "default"
-        ? this.moment(value)
-        : (this.moment as any).tz(value, timezone);
-
-    if (this.locale === undefined && locale === undefined) {
-      return parsedValue;
-    }
-    return parsedValue.locale(locale ?? this.locale);
+    return timezone === "default"
+      ? this.dayjs(value).locale(locale ?? this.locale)
+      : this.dayjs.tz(value, timezone).locale(locale ?? this.locale);
   };
 
   /**
-   * Creates a Moment.js date object from a string or returns an invalid date.
+   * Creates a Day.js date object from a string or returns an invalid date.
    * @param value - The date string to parse.
    * @param timezone - The timezone to use (default is "default").
    * @param locale - The locale to use for parsing.
-   * @returns The parsed Moment.js date object or an invalid date object.
+   * @returns The parsed Day.js date object or an invalid date object.
    */
   public date = <T extends string | undefined>(
     value?: T,
     timezone: Timezone = "default",
     locale?: string,
-  ): Moment => {
+  ): Dayjs => {
     if (!value || !this.isValidDateString(value)) {
-      return this.moment.invalid();
+      return this.dayjs("Invalid Date");
     }
-
+    let date: Dayjs;
     if (timezone === "UTC") {
-      return this.createUTCDate(value, locale);
+      date = this.createUTCDate(value, locale);
+    } else if (timezone === "system" || timezone === "default") {
+      date = this.createSystemDate(value, locale);
+    } else {
+      date = this.createTZDate(value, timezone, locale);
     }
 
-    if (
-      timezone === "system" ||
-      (timezone === "default" && !this.hasTimezonePlugin())
-    ) {
-      return this.createSystemDate(value, locale);
-    }
-
-    return this.createTZDate(value, timezone, locale);
+    return date;
   };
 
   /**
-   * Formats a Moment.js date object using the specified format string.
+   * Formats a Day.js date object using the specified format string.
    * Returns an empty string when null or undefined date is given.
-   * @param date - The Moment.js date object to format.
+   * @param date - The Day.js date object to format.
    * @param format - The format string to use.
    * @param locale - The locale to use for formatting.
    * @returns The formatted date string.
    */
   public format(
-    date: Moment | null | undefined,
+    date: Dayjs | null | undefined,
     format: RecommendedFormats = "DD MMM YYYY",
     locale?: string,
   ): string {
     if (this.isValid(date)) {
-      return date
-        .clone()
-        .locale(locale ?? this.locale)
-        .format(format);
+      return date.locale(locale ?? this.locale).format(format);
     }
     return "";
   }
 
   /**
-   * Compares two Moment.js date objects.
-   * @param dateA - The first Moment.js date object.
-   * @param dateB - The second Moment.js date object.
+   * Compares two Day.js date objects.
+   * @param dateA - The first Day.js date object.
+   * @param dateB - The second Day.js date object.
    * @returns 0 if equal, 1 if dateA is after dateB, -1 if dateA is before dateB.
    */
-  public compare(dateA: Moment, dateB: Moment): number {
+  public compare(dateA: Dayjs, dateB: Dayjs): number {
     if (dateA.isSame(dateB)) {
       return 0;
     }
@@ -196,11 +202,13 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
     value: string,
     format: string,
     locale?: string,
-  ): ParserResult<Moment> {
-    const parsedDate =
-      this.locale || locale
-        ? this.moment(value, format, locale ?? this.locale, true)
-        : this.moment(value, format, true);
+  ): ParserResult<Dayjs> {
+    const parsedDate = this.dayjs(
+      value?.trim(),
+      format,
+      locale ?? this.locale,
+      false,
+    );
     if (parsedDate.isValid()) {
       return {
         date: parsedDate,
@@ -209,36 +217,36 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
     }
     const isDateDefined = !!value?.trim().length;
     return {
-      date: parsedDate,
+      date: this.dayjs("Invalid Date"),
       value,
       errors: [
         {
           message: isDateDefined ? "not a valid date" : "no date defined",
           type: isDateDefined
-            ? DateDetailErrorEnum.INVALID_DATE
-            : DateDetailErrorEnum.UNSET,
+            ? DateDetailError.INVALID_DATE
+            : DateDetailError.UNSET,
         },
       ],
     };
   }
 
   /**
-   * Checks if a Moment.js date object is valid.
-   * @param date - The Moment.js date object to check.
+   * Checks if a Day.js date object is valid.
+   * @param date - The Day.js date object to check.
    * @returns True if the date is valid date object, false otherwise.
    */
-  public isValid(date: any): date is Moment {
-    return this.moment.isMoment(date) ? date.isValid() : false;
+  public isValid(date: any): date is Dayjs {
+    return this.isDayjs(date) ? date.isValid() : false;
   }
 
   /**
-   * Subtracts time from a Moment.js date object.
-   * @param date - The Moment.js date object to subtract from.
+   * Subtracts time from a Day.js date object.
+   * @param date - The Day.js date object to subtract from.
    * @param duration - The duration to subtract.
-   * @returns The resulting Moment.js date object.
+   * @returns The resulting Day.js date object.
    */
   public subtract(
-    date: Moment,
+    date: Dayjs,
     {
       days,
       weeks,
@@ -258,19 +266,19 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
       seconds?: number;
       milliseconds?: number;
     },
-  ): Moment {
-    let newDate = date.clone();
+  ): Dayjs {
+    let newDate = date;
     if (days) {
-      newDate = newDate.subtract(days, "days");
+      newDate = newDate.subtract(days, "day");
     }
     if (weeks) {
-      newDate = newDate.subtract(weeks, "weeks");
+      newDate = newDate.subtract(weeks, "week");
     }
     if (months) {
-      newDate = newDate.subtract(months, "months");
+      newDate = newDate.subtract(months, "month");
     }
     if (years) {
-      newDate = newDate.subtract(years, "years");
+      newDate = newDate.subtract(years, "year");
     }
     if (hours) {
       newDate = newDate.subtract(hours, "hours");
@@ -288,13 +296,13 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
   }
 
   /**
-   * Adds time to a Moment.js date object.
-   * @param date - The Moment.js date object to add to.
+   * Adds time to a Day.js date object.
+   * @param date - The Day.js date object to add to.
    * @param duration - The duration to add.
-   * @returns The resulting Moment.js date object.
+   * @returns The resulting Day.js date object.
    */
   public add(
-    date: Moment,
+    date: Dayjs,
     {
       days,
       weeks,
@@ -314,19 +322,19 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
       seconds?: number;
       milliseconds?: number;
     },
-  ): Moment {
-    let newDate = date.clone();
+  ): Dayjs {
+    let newDate = date;
     if (days) {
-      newDate = newDate.add(days, "days");
+      newDate = newDate.add(days, "day");
     }
     if (weeks) {
-      newDate = newDate.add(weeks, "weeks");
+      newDate = newDate.add(weeks, "week");
     }
     if (months) {
-      newDate = newDate.add(months, "months");
+      newDate = newDate.add(months, "month");
     }
     if (years) {
-      newDate = newDate.add(years, "years");
+      newDate = newDate.add(years, "year");
     }
     if (hours) {
       newDate = newDate.add(hours, "hour");
@@ -344,13 +352,13 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
   }
 
   /**
-   * Sets specific components of a Moment.js date object.
-   * @param date - The Moment.js date object to modify.
+   * Sets specific components of a Day.js date object.
+   * @param date - The Day.js date object to modify.
    * @param components - The components to set, the month is a number (1-12).
-   * @returns The resulting Moment.js date object.
+   * @returns The resulting Day.js date object.
    */
   public set(
-    date: Moment,
+    date: Dayjs,
     {
       day,
       month,
@@ -368,7 +376,7 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
       second?: number;
       millisecond?: number;
     },
-  ): Moment {
+  ): Dayjs {
     let newDate = date.clone();
     if (day) {
       newDate = newDate.date(day);
@@ -395,50 +403,48 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
   }
 
   /**
-   * Checks if two Moment.js date objects are the same based on the specified granularity.
-   * @param dateA - The first Moment.js date object.
-   * @param dateB - The second Moment.js date object.
+   * Checks if two Day.js date objects are the same based on the specified granularity.
+   * @param dateA - The first Day.js date object.
+   * @param dateB - The second Day.js date object.
    * @param granularity - The granularity to compare by ("day", "month", "year").
    * @returns True if the dates are the same, false otherwise.
    */
   public isSame(
-    dateA: Moment,
-    dateB: Moment,
+    dateA: Dayjs,
+    dateB: Dayjs,
     granularity: "day" | "month" | "year" = "day",
   ): boolean {
     return dateA.isSame(dateB, granularity);
   }
 
   /**
-   * Gets the start of a specified time period for a Moment.js date object.
-   * @param date - The Moment.js date object.
+   * Gets the start of a specified time period for a Day.js date object.
+   * @param date - The Day.js date object.
    * @param offset - The time period ("day", "week", "month", "year").
    * @param locale - The locale to use.
-   * @returns The Moment.js date object representing the start of the period.
+   * @returns The Day.js date object representing the start of the period.
    */
   public startOf(
-    date: Moment,
+    date: Dayjs,
     offset: "day" | "week" | "month" | "year",
     locale?: string,
-  ): Moment {
-    const newDate = date.clone().locale(locale ?? this.locale);
-    return newDate.startOf(offset);
+  ): Dayjs {
+    return date.locale(locale ?? this.locale).startOf(offset);
   }
 
   /**
-   * Gets the end of a specified time period for a Moment.js date object.
-   * @param date - The Moment.js date object.
+   * Gets the end of a specified time period for a Day.js date object.
+   * @param date - The Day.js date object.
    * @param offset - The time period ("day", "week", "month", "year").
    * @param locale - The locale to use.
-   * @returns The Moment.js date object representing the end of the period.
+   * @returns The Day.js date object representing the end of the period.
    */
   public endOf(
-    date: Moment,
+    date: Dayjs,
     offset: "day" | "week" | "month" | "year",
     locale?: string,
-  ): Moment {
-    const newDate = date.clone().locale(locale ?? this.locale);
-    return newDate.endOf(offset);
+  ): Dayjs {
+    return date.locale(locale ?? this.locale).endOf(offset);
   }
 
   /**
@@ -446,8 +452,8 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
    * @param locale - The locale to use.
    * @returns The current date at the start of the day.
    */
-  public today(locale?: string): Moment {
-    return this.moment()
+  public today(locale?: string): Dayjs {
+    return this.dayjs()
       .locale(locale ?? this.locale)
       .startOf("day");
   }
@@ -457,15 +463,25 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
    * @param locale - The locale to use.
    * @returns The current date and time.
    */
-  public now(locale?: string): Moment {
-    return this.moment().locale(locale ?? this.locale);
+  public now(locale?: string): Dayjs {
+    return this.dayjs().locale(locale ?? this.locale);
+  }
+
+  /**
+   * Gets the day of the week for a Day.js date object.
+   * @param date - The Day.js date object.
+   * @param locale - The locale to use.
+   * @returns The day of the week as a number (0-6).
+   */
+  public getDayOfWeek(date: Dayjs, locale?: string): number {
+    return date.locale(locale ?? this.locale).day();
   }
 
   /**
    * Gets the name of the day of the week.
    * @param dow - The day of the week as a number (0-6).
    * @param format - The format for the day name ("long", "short", "narrow").
-   * @param locale - The locale to use.
+   * @param locale - The locale to use
    * @returns The name of the day of the week.
    */
   public getDayOfWeekName(
@@ -473,59 +489,52 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
     format: "long" | "short" | "narrow",
     locale?: string,
   ): string {
-    const day = this.moment()
+    const date = this.dayjs()
       .locale(locale ?? this.locale)
       .weekday(dow);
-    return format === "narrow" ? day.format("dd")[0] : day.format("dddd");
+    const formatString =
+      format === "long" ? "dddd" : format === "short" ? "ddd" : "dd";
+    return date.format(formatString);
   }
 
   /**
-   * Gets the day of the week for a Moment.js date object.
-   * @param date - The Moment.js date object.
-   * @returns The day of the week as a number (0-6).
-   */
-  public getDayOfWeek(date: Moment): number {
-    return date.day();
-  }
-
-  /**
-   * Gets the day of the month for a Moment.js date object.
-   * @param date - The Moment.js date object.
+   * Gets the day of the month for a Day.js date object.
+   * @param date - The Day.js date object.
    * @returns The day of the month as a number (1-31).
    */
-  public getDay(date: Moment): number {
+  public getDay(date: Dayjs): number {
     return date.date();
   }
 
   /**
-   * Gets the month for a Moment.js date object.
-   * @param date - The Moment.js date object.
-   * @returns The month as a number (1-12).
+   * Gets the month for a Day.js date object.
+   * @param date - The Day.js date object.
+   * @returns The month as a number (0-11).
    */
-  public getMonth(date: Moment): number {
+  public getMonth(date: Dayjs): number {
     return date.month() + 1;
   }
 
   /**
-   * Gets the year for a Moment.js date object.
-   * @param date - The Moment.js date object.
+   * Gets the year for a Day.js date object.
+   * @param date - The Day.js date object.
    * @returns The year as a number.
    */
-  public getYear(date: Moment): number {
+  public getYear(date: Dayjs): number {
     return date.year();
   }
 
   /**
-   * Gets the time components for a Moment.js date object.
-   * @param date - The Moment.js date object.
+   * Gets the time components for a Day.js date object.
+   * @param date - The Day.js date object.
    * @returns An object containing the hour, minute, second, and millisecond.
    */
-  public getTime(date: Moment): TimeFields {
+  public getTime(date: Dayjs): TimeFields {
     return {
-      hour: date.hour(),
-      minute: date.minute(),
-      second: date.second(),
-      millisecond: date.millisecond(),
+      hour: date.get("hour"),
+      minute: date.get("minute"),
+      second: date.get("second"),
+      millisecond: date.get("millisecond"),
     };
   }
 
@@ -534,15 +543,16 @@ export class AdapterMoment implements SaltDateAdapter<Moment, string> {
    * @param value
    */
   public isValidDateString(value: string): boolean {
-    /** Ensure ISO 8601 format of date string is passed to Moment to avoid warning **/
-    return this.moment(value, this.moment.ISO_8601, true).isValid();
+    // Attempt to parse the string as an ISO 8601 date
+    const date = this.dayjs(value, { format: "YYYY-MM-DDTHH:mm:ssZ" });
+    return date.isValid();
   }
 
   /**
    * Clone the date object
    * @param date
    */
-  public clone(date: Moment): Moment {
+  public clone(date: Dayjs): Dayjs {
     return date.clone();
   }
 }
