+++ packages/date-adapters/src/types/DateFrameworkTypeMap.ts
@@ -1,0 +1,12 @@
+/**
+ * To add a new adapter, then, add the adapter's date object to `DateFrameworkTypeMap` interface
+ *
+ * declare module "@salt-ds/date-adapters" {
+ *   interface DateFrameworkTypeMap {
+ *     luxon: DateTime;
+ *   }
+ * }
+ */
+
+// biome-ignore lint/complexity/noBannedTypes: type augmented by configured adapters
+export type DateFrameworkTypeMap = {};
