+++ packages/lab/src/__tests__/__e2e__/calendar/Calendar.offset.cy.tsx
@@ -1,11 +1,11 @@
-import type {
-  DateFrameworkType,
-  SaltDateAdapter,
+import {
+  AdapterDateFns,
+  AdapterDayjs,
+  AdapterLuxon,
+  AdapterMoment,
+  type DateFrameworkType,
+  type SaltDateAdapter,
 } from "@salt-ds/date-adapters";
-import { AdapterDateFns } from "@salt-ds/date-adapters";
-import { AdapterDayjs } from "@salt-ds/date-adapters";
-import { AdapterLuxon } from "@salt-ds/date-adapters";
-import { AdapterMoment } from "@salt-ds/date-adapters";
 import {
   Calendar,
   CalendarGrid,
