+++ site/src/examples/localization-provider/Locale.tsx
@@ -4,6 +4,10 @@ import {
   RadioButton,
   RadioButtonGroup,
 } from "@salt-ds/core";
+import { AdapterDateFns } from "@salt-ds/date-adapters/date-fns";
+import { AdapterDayjs } from "@salt-ds/date-adapters/dayjs";
+import { AdapterLuxon } from "@salt-ds/date-adapters/luxon";
+import { AdapterMoment } from "@salt-ds/date-adapters/moment";
 import {
   Calendar,
   CalendarGrid,
@@ -13,10 +17,6 @@ import {
 } from "@salt-ds/lab";
 import { type ChangeEventHandler, type ReactElement, useState } from "react";
 import "dayjs/locale/en";
-import { AdapterDateFns } from "@salt-ds/date-adapters";
-import { AdapterDayjs } from "@salt-ds/date-adapters";
-import { AdapterLuxon } from "@salt-ds/date-adapters";
-import { AdapterMoment } from "@salt-ds/date-adapters";
 import moment from "moment";
 import "moment/locale/zh-cn"; // Import the Chinese locale
 import { enUS as dateFnsEnUs } from "date-fns/locale";
