+++ site/tsconfig.json
@@ -3,7 +3,10 @@
   "compilerOptions": {
     "baseUrl": ".",
     "jsx": "preserve",
-    "incremental": true
+    "incremental": true,
+    "paths": {
+      "@salt-ds/date-adapters/*": ["../packages/date-adapters/src/index.ts"]
+    }
   },
   "include": ["src/**/*", "globals.d.ts"],
   "exclude": ["node_modules"]
