+++ vite.config.ts
@@ -1,6 +1,8 @@
+import tsconfigPaths from "vite-tsconfig-paths";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
+  plugins: [tsconfigPaths() as any],
   test: {
     include: ["**/*.spec.[jt]s?(x)"],
   },
